refactor: notification endpoints now self-register
