+++ src/flows/pipes/Notification/endpoints/Slack/view.tsx
